requirements.txt
tensorflow==2.3.0
pandas==1.1.3
matplotlib==3.3.2
numpy==1.18.5
streamlit==0.78.0
Keras==2.4.3
Pillow==8.1.2
scikit_learn==0.24.1
